Fix ItemTest
