+++ common/src/test/scala/uk/ac/wellcome/models/IdentifiedItemTest.scala
@@ -5,45 +5,12 @@ import uk.ac.wellcome.utils.JsonUtil._
 import uk.ac.wellcome.test.utils.JsonTestUtil
 import uk.ac.wellcome.utils.JsonUtil
 
-class ItemTest extends FunSpec with Matchers with JsonTestUtil {
+class IdentifiedItemTest extends FunSpec with Matchers with JsonTestUtil {
 
   val identifiedItemJson: String =
     s"""
       |{
       |  "canonicalId": "canonicalId",
-      |  "sourceIdentifier": {
-      |      "identifierScheme": "${IdentifierSchemes.miroImageNumber.toString}",
-      |      "value": "value"
-      |  },
-      |  "identifiers": [
-      |    {
-      |      "identifierScheme": "${IdentifierSchemes.miroImageNumber.toString}",
-      |      "value": "value"
-      |    }
-      |  ],
-      |  "locations": [
-      |    {
-      |      "locationType": "location",
-      |      "url" : "",
-      |      "credit" : null,
-      |      "license": {
-      |        "licenseType": "${License_CCBY.licenseType}",
-      |        "label": "${License_CCBY.label}",
-      |        "url": "${License_CCBY.url}",
-      |        "type": "License"
-      |      },
-      |      "type": "DigitalLocation"
-      |    }
-      |  ],
-      |  "visible":true,
-      |  "type": "Item"
-      |}
-    """.stripMargin
-
-  val unidentifiedItemJson: String =
-    s"""
-      |{
-      |  "canonicalId" : null,
       |  "sourceIdentifier": {
       |      "identifierScheme": "${IdentifierSchemes.miroImageNumber.toString}",
       |      "value": "value"
@@ -84,33 +51,12 @@ class ItemTest extends FunSpec with Matchers with JsonTestUtil {
     value = "value"
   )
 
-  val unidentifiedItem = Item(
-    canonicalId = None,
+  val identifiedItem = IdentifiedItem(
+    canonicalId = "canonicalId",
     sourceIdentifier = identifier,
     identifiers = List(identifier),
     locations = List(location)
   )
-
-  val identifiedItem = Item(
-    canonicalId = Some("canonicalId"),
-    sourceIdentifier = identifier,
-    identifiers = List(identifier),
-    locations = List(location)
-  )
-
-  it("should serialise an unidentified Item as JSON") {
-    val result = JsonUtil.toJson(unidentifiedItem)
-
-    result.isSuccess shouldBe true
-    assertJsonStringsAreEqual(result.get, unidentifiedItemJson)
-  }
-
-  it("should deserialize a JSON string as a unidentified Item") {
-    val result = JsonUtil.fromJson[Item](unidentifiedItemJson)
-
-    result.isSuccess shouldBe true
-    result.get shouldBe unidentifiedItem
-  }
 
   it("should serialise an identified Item as JSON") {
     val result = JsonUtil.toJson(identifiedItem)
